requirements.txt
gpiozero
paho-mqtt
pigpio
python-dotenv
